feat(context/auth): update use of AuthContextProvider and useAuth
