--- src/hooks/use-auth/index.ts
@@ -1,1 +1,0 @@
-export * from './use-auth'
