--- src/hooks/use-auth/use-auth.types.ts
@@ -1,10 +1,0 @@
-export type SignInCredentials = {
-  email: string
-  password: string
-}
-
-export type SignUpCredentials = {
-  email: string
-  password: string
-  username: string
-}
